requirements.txt
ultralytics
opencv-python
torch
torchvision
numpy
matplotlib
pillow
seaborn
pandas